gc: update gc API to improve usage and fix bug (#340)
C compilers use different amount of padding in the cortecs_array structs based on the alignment of the elements array type. calls to the gc array alloc need to account for this. We dont want users of the API to be concerned with this detail. This is a small redesign of the GC api to avoid users of the API from always having to submit finalizer and allocation details to the API while accounting for the varying alignment/padding