gc: spoof allocation logging message during gc init
A couple of GC logging messages get lost during initialization because of circular dependencies between the log being initialized to be able to log, but the GC being already initialized to be able to allocate the log. This leads to 2 alloc log messages and 2 enqueue_dec log messages being lost. Spoof these messages for log completeness